requirements.txt
boto3>=1.18.0
redis>=3.5.3
grpcio>=1.38.1
protobuf>=3.17.3
PyYAML>=5.4.1
